fix: move dereference pipeline to utils file
