+++ tests/unit/dereference-pipeline.spec.ts
@@ -1,5 +1,5 @@
 import { Pipeline } from '@/lib/steps';
-import { dereferencePipelines, PipelinesScopeContext } from '@/store/state';
+import { dereferencePipelines, PipelinesScopeContext } from '@/store/utils/dereference-pipeline';
 
 describe('dereferencePipelines', () => {
   it('should let unmodified pipeline that are not referenced', () => {
